% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.r
\name{addDoubletScores}
\alias{addDoubletScores}
\title{addDoubletScores}
\usage{
addDoubletScores(obj, dbobj)
}
\arguments{
\item{obj}{An amethyst object containing the original cell data.}

\item{dbobj}{An amethyst object containing the doublet scores in its metadata.}
}
\value{
The original amethyst object with updated metadata including doublet scores.
}
\description{
This function adds doublet scores from a doublet object to the metadata of the original amethyst object.
}
\examples{
\dontrun{
  obj <- addDoubletScores(obj, dbobj)
}
}
