% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{addIndex}
\alias{addIndex}
\title{Get indexes for a new gene and append it to the pre-existing index slot}
\usage{
addIndex(obj, type, threads = 1, genes)
}
\arguments{
\item{obj}{Amethyst object to append an index}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG. Can be a list.}

\item{threads}{Optional threads parameter to use if parallelizing. If running locally, use with caution.}

\item{genes}{List of genes to index.}
}
\description{
Get indexes for a new gene and append it to the pre-existing index slot
}
