% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.r
\name{addMatrixMeans}
\alias{addMatrixMeans}
\title{addMatrixMeans}
\usage{
addMatrixMeans(obj, matrix_name, name)
}
\arguments{
\item{obj}{An amethyst object containing the original cell data.}

\item{matrix_name}{A character string specifying the name of the genome matrix to process.}

\item{name}{A character string specifying the name of the new column to be added to the metadata.}
}
\value{
An amethyst object with updated metadata including the mean values from the specified genome matrix.
}
\description{
This function calculates the mean values for each cell (column) in the specified genome matrix, excluding NAs,
and adds these mean values as a new column in the metadata of the provided amethyst object.
If you want to process irlba where mean is used for the NA values, this function can be used.
}
