% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addPrefix}
\alias{addPrefix}
\title{addPrefix}
\usage{
addPrefix(obj, prefix)
}
\arguments{
\item{obj}{Name of object to add prefixes to}

\item{prefix}{Name of prefix to append, e.g. "A_"}
}
\value{
Returns an updated amethyst object containing the prefix
}
\description{
Helper function for adding a barcode prefix to appropriate locations throughout the object.
At this time, only one prefix can be added per object.
}
\examples{
\dontrun{
 new_obj <- addPrefix(obj, prefix = "A_")
}
}
