% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-Lauren’s MacBook Pro.R, R/helper.R
\name{aggregateMatrix}
\alias{aggregateMatrix}
\title{aggregateMatrix}
\usage{
aggregateMatrix(obj, matrix, groupBy)

aggregateMatrix(obj, matrix, groupBy)
}
\arguments{
\item{obj}{Amethyst object containing the matrix to be aggregated}

\item{matrix}{Name of matrix contained in the genomeMatrices slot to aggregate}

\item{groupBy}{Parameter in the metadata to aggregate over}
}
\value{
Returns a matrix containing the mean value per group

Returns a matrix containing the mean value per group
}
\description{
Aggregate genomic window matrices by a categorical variable contained in the metadata

Aggregate genomic window matrices by a categorical variable contained in the metadata
}
\examples{
obj <- aggregateMatrix(obj, matrix = "ch_100k_pct", groupBy = cluster_id, name = "cluster_ch_100k_pct")
obj <- aggregateMatrix(obj, matrix = "ch_100k_pct", groupBy = cluster_id, name = "cluster_ch_100k_pct")
}
