% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.r
\name{buildDoubletModel}
\alias{buildDoubletModel}
\title{buildDoubletModel}
\usage{
buildDoubletModel(dbobj, method = "rf", reduction = "irlba")
}
\arguments{
\item{dbobj}{An amethyst object containing the data with doublets.}

\item{method}{A character string specifying the machine learning method to use for model training. Default is 'rf' (random forest).}

\item{reduction}{A character string specifying the name of the reduction slot to use. Default is 'irlba'.}
}
\value{
A list containing the trained model, the predictions on the test set, and the test data.
}
\description{
This function builds a model to identify doublets using a specified machine learning method.
The function splits the data into training and test sets, trains the model, and predicts probabilities on the test set.
}
\examples{
\dontrun{
result <- buildDoubletModel(dbobj, reduction = "irlba", method="rf")
}
}
