% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{buildMetacellH5Index}
\alias{buildMetacellH5Index}
\title{buildMetacellH5Index}
\usage{
buildMetacellH5Index(h5file, type = "CG", use_cache = TRUE)
}
\arguments{
\item{h5file}{Path to metacell H5 (e.g. "metacell_cg_1kbp.h5").}

\item{type}{H5 group name (e.g. "CG").}

\item{use_cache}{Logical; reuse cached index if available (default TRUE).}
}
\value{
data.table with columns: idx, chr, start, end (keyed by chr,start,end).
}
\description{
Build or retrieve a cached index for a metacell H5 file.
}
