% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{calcSmoothedWindows}
\alias{calcSmoothedWindows}
\title{calcSmoothedWindows}
\usage{
calcSmoothedWindows(
  obj,
  type = "CG",
  threads = 1,
  step = 500,
  smooth = 3,
  genome = "hg38",
  subset = NULL,
  index = "chr_cg",
  futureType = "multicore",
  groupBy = "cluster_id",
  returnSumMatrix = TRUE,
  returnPctMatrix = TRUE,
  save = FALSE,
  chrList = NULL,
  chrSizes = NULL
)
}
\arguments{
\item{obj}{Amethyst object containing the h5 file paths}

\item{type}{Type of methylation - e.g. "CG" or "CH" - to calculate. Typically will be "CG"}

\item{threads}{Optional number of threads for parallelization (multithreading strongly recommended)}

\item{step}{Width of the genomic window to calculate. Default is 500 bp}

\item{smooth}{Number of windows to include surrounding the target region; i.e. produces a sliding window matrix
Default parameter is 3, resulting in a 1500 x 500 bp sliding window matrix}

\item{genome}{Genome build of the organism(s) being analyzed. Options are currently "hg19", "hg38", "mm10", or "mm39".}

\item{subset}{Optional; only calculate for portion of genome. Specify ranges like: "chr15_77563027_77870900"}

\item{index}{Name of chr index in the index slot}

\item{futureType}{If using parallelization, should R multithread using "multicore" or "multisession"}

\item{groupBy}{Parameter contained in the metadata over which to aggregate observations. Default is "cluster_id"}

\item{returnSumMatrix}{Whether or not the function should return the matrix of summed c and t observations. Required for testDMR input}

\item{returnPctMatrix}{Whether or not the function should calculate \% methylation over each genomic window. Required for heatMap input}

\item{save}{Optional; save each chromosome output intermediate. Helpful if you are working with large datasets}

\item{chrList}{Optional; chromosome whitelist (only use if necessary)}

\item{chrSizes}{Optional; matched size vector of chromosomes provided in the whitelist (MUST include if chrList is provided)}
}
\value{
if returnSumMatrix = TRUE, returns a data.table with the genomic location as chr, start, and end columns,
plus aggregated c and t observations for each groupBy value. If returnPctMatrix = TRUE, returns a data.table with the
genomic location as chr, start, and end columns, plus one column of the mean \% methylation for each window per groupBy value.
If both are true, returns a list containing each result.
}
\description{
Aggregate c and t observations over smoothed genomic windows of a user-defined size.
Optionally calculate the \% methylation for each window.
}
\examples{
\dontrun{
  output <- calcSmoothedWindows(obj, groupBy = "cluster_id")
  output <- calcSmoothedWindows(obj, groupBy = "louvain_vmr_cluster", chrList = c("chr1", "chr2"), chrSizes = c(248956422, 242193529))
}
}
