% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calcWindowMethylationStat}
\alias{calcWindowMethylationStat}
\title{Calculate Window Methylation Statistic}
\usage{
calcWindowMethylationStat(h5path, barcode, stepsize, stat)
}
\arguments{
\item{h5path}{Path to the .h5 file containing the methylation data for each
cell}

\item{barcode}{The barcode for a specific cell within the .h5 file}

\item{stat}{One of \code{'percent'}, \code{'ration'} or \code{'score'}; the statistic to
calculate for the window}

\item{stepszie}{The size of the windows}
}
\value{
Data.table of the cell methylation data updated to include the stat
}
\description{
For a given genomic region, calculates a specific statistic
(such as percent methylated).
}
