% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{clusterCompare}
\alias{clusterCompare}
\title{Cluster compare}
\usage{
clusterCompare(
  matrix,
  ref = celltyperefs,
  level,
  type,
  method = "pearson",
  n = 5,
  aggregate = NULL
)
}
\arguments{
\item{matrix}{Average percent methylation aggregated over 100kb windows by a grouping factor (see makeWindows). Column names should be genomic positions.}

\item{ref}{The reference structure to correlate to. The default for brain used here is from PMC8494641.}

\item{level}{The grouping level in the reference to compare to}

\item{type}{The type of methylation in the reference to compare to (cg or ch). Make sure the input matrix reflects the same type.}

\item{method}{Correlation coefficient to be computed ("pearson", "kendall", or "spearman")}

\item{n}{Number of top correlations to plot.}
}
\value{
A ggplot dot plot showing the top 5 correlations of each group to the reference
}
\description{
Correlates average percent methylation over 100kb windows aggregated by metadata to a reference matrix to aid in cell type identification.
}
\examples{
clusterCompare(cluster_100kb_cg, level = "major_type", type = "cg")
}
