% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{clusterLabelTransfer}
\alias{clusterLabelTransfer}
\title{clusterLabelTransfer}
\usage{
clusterLabelTransfer(obj, labels = "type", cluster_ids = "cluster_id")
}
\arguments{
\item{obj}{An Amethyst object to perform cluster label transfer on. The object must contain metadata with the specified labels and cluster IDs.}

\item{labels}{A character string specifying the metadata column in the Amethyst object that contains the labels to be transferred. Defaults to 'type'.}

\item{cluster_ids}{A character string specifying the metadata column in the Amethyst object that contains the cluster IDs. Defaults to 'cluster_id'.}
}
\value{
The Amethyst object with updated metadata containing the transferred cluster labels.
}
\description{
Transfers cluster labels from one set of annotations to another within an Amethyst object.
This function calculates the most common label within each cluster and assigns it as the representative label for that cluster.
}
