% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{collapseDMR}
\alias{collapseDMR}
\title{collapseDMR}
\usage{
collapseDMR(
  obj,
  dmrMatrix,
  maxDist = 0,
  minLength = 1000,
  reduce = TRUE,
  mergeDirections = FALSE,
  annotate = TRUE
)
}
\arguments{
\item{obj}{Amethyst object containing the h5 file paths}

\item{dmrMatrix}{Table of differentially methylated regions identified with testDMR and filterDMR}

\item{maxDist}{Numeric; Maximum allowable overlap between DMRs to be considered adjacent}

\item{minLength}{Minimum length of collapsed DMR window to include in the output}

\item{reduce}{Boolean; Reduce results to unique observations (recommended)}

\item{mergeDirections}{Boolean; If TRUE, merges adjacent DMRs regardless of direction (hypo/hypermethylated).
Recommended to keep FALSE. Can be useful for annotating large loci of interest}

\item{annotate}{Boolean; Add column with overlapping gene names. Requires annotation file in @ref slot}
}
\value{
Returns a data.table of collapsed DMRs
}
\description{
Identify adjacent DMRs
}
\examples{
\dontrun{
  results <- collapseDMR(dmrs, maxDist = 2000, minLength = 2000, reduce = T, annotate = T)
}
}
