% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{combineObject}
\alias{combineObject}
\title{combineObject}
\usage{
combineObject(objList, genomeMatrices)
}
\arguments{
\item{objList}{List of amethyst objects to merge. Make sure barcode names are unique.}

\item{matrices}{List of genomeMatrices contained within each amethyst object to merge.}
}
\value{
Returns a new object with merged slots. Dimensionality reductions should be re-run post merging.
}
\description{
Merge a list of amethyst objects into one
}
\examples{
new <- combineObject(objList = list(obj1, obj2, obj3), genomeMatrices = c("ch_100k_pct", "gene_ch"))
}
