% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convertObject}
\alias{convertObject}
\title{convertObject}
\usage{
convertObject(obj)
}
\arguments{
\item{obj}{Old amethyst object to convert}
}
\value{
Returns an updated amethyst object with slots:
h5paths, genomeMatrices, tracks, reductions, index, metadata, ref, results
}
\description{
Shortcut for converting old amethyst object format to new.
Adds "tracks" and "results" slots.
}
