% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{createScaleObject}
\alias{createScaleObject}
\title{createScaleObject}
\usage{
createScaleObject(directory, genomeMatrices = NULL)
}
\arguments{
\item{directory}{Path to the directory containing output from the Scale Biosciences computational pipeline.
The folder is expected to contain an ".allCells.csv" file with metadata, a genome_bin_matrix folder with pre-constructed
matrices, and a methylation_coverage folder containing .h5 files with base-level methylation information for each cell.}

\item{genomeMatrices}{Optional name of pre-constructed matrices in the genome_bin_matrix folder to include.}
}
\value{
Returns a populated amethyst object for futher analysis.
}
\description{
Helper function for converting Scale Biosciences pipeline output into an Amethyst object
}
\examples{
obj <- createScaleObject(directory = "~/Downloads/scalebio", genomeMatrices = c("CG.score"))
}
