% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{dimFeature}
\alias{dimFeature}
\title{dimFeature}
\usage{
dimFeature(obj, colorBy, colors = NULL, pointSize = 0.5, reduction = "umap")
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{colorBy}{The metadata feature to plot in UMAP or TSNE space}

\item{colors}{Optional color scale}

\item{pointSize}{Optional adjustment of point size}

\item{reduction}{Whether to plot "umap" or "tsne" coordinates. Note runUmap and/or runTsne must have been run.}
}
\value{
Returns a ggplot graph plotting xy coordinates of cells colored according to the specified feature
}
\description{
Plot any feature in the metadata in UMAP or TSNE space
}
\examples{
dimFeature(obj = obj, colorBy = "cluster_id", reduction = "umap")
}
