% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{dimM}
\alias{dimM}
\title{dimM}
\usage{
dimM(
  obj,
  genes,
  matrix,
  blend = FALSE,
  colors = NULL,
  squish = NULL,
  reduction = "umap",
  pointSize = 0.1,
  nrow = round(sqrt(length(genes)))
)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{genes}{List of genes to plot \% methylation in UMAP or TSNE space
Theoretically, this could also be the name of a non-gene feature, as long as they are rownames
in your specified matrix}

\item{matrix}{Matrix in the genomeMatrices slot containing aggregated methylation levels over features per cell}

\item{blend}{Boolean indicating whether to blend methylation levels of two or three genes}

\item{colors}{Optional; color gradient to include}

\item{squish}{Optional; upper limit of the color scale. Everything above this will be plotted as max}

\item{reduction}{Name of reduction slot containing UMAP / TSNE dimensionality reduction coordinates}

\item{pointSize}{Optional; adjust point size}

\item{nrow}{Number of rows to distribute graphs over}
}
\value{
Returns a plot of methylation over given genomic region over UMAP or TSNE coordinates
}
\description{
Plot cumulative \% methylation of a feature over UMAP / TSNE coordinates
}
\examples{
\dontrun{
  dimM(obj, genes = c("GRM8", "SATB1"), matrix = "gene_ch", pointSize = 0.8)
  dimM(obj, genes = c("GRM8", "SATB1"), blend = T, matrix = "gene_ch_imputed", pointSize = 0.8)
}
}
