% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{.get_ref_dt_and_chr_col}
\alias{.get_ref_dt_and_chr_col}
\title{.get_ref_dt_and_chr_col}
\usage{
.get_ref_dt_and_chr_col(obj)
}
\description{
Internal helper to coerce obj@ref to data.table and detect chromosome column.
}
\keyword{internal}
