% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{dotGeneM}
\alias{dotGeneM}
\title{dotGeneM}
\usage{
dotGeneM(obj, genes, groupBy, matrix = NULL, type = NULL, metric = NULL)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{genes}{List of genes to plot methylation on the x axis}

\item{groupBy}{A categorical variable contained in the metadata to group cells by on the y axis}

\item{matrix}{Optional name of a pre-calculated matrix of values contained in the genomeMatrices slot to use}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG.}

\item{metric}{If matrix is not specified, what methylation parameter to calculate, i.e. "percent", "score", or "ratio"}
}
\value{
Returns a ggplot object displaying average methylation values for each gene by the grouping variable
}
\description{
Generates a dot plot of percent methylation over a gene body
}
\examples{
dotGeneM(obj = obj, genes = c("SATB2", "GAD1"), type = "gene_ch", groupBy = "cluster_id")
}
