% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{extractAttributes}
\alias{extractAttributes}
\title{extractAttributes from https://www.biostars.org/p/272889/
Extract information from the attributes column of a gtf file}
\usage{
extractAttributes(gtf_attributes, att_of_interest)
}
\arguments{
\item{att_of_interest}{}
}
\description{
extractAttributes from https://www.biostars.org/p/272889/
Extract information from the attributes column of a gtf file
}
