% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{filterDMR}
\alias{filterDMR}
\title{filterDMR}
\usage{
filterDMR(
  dmrMatrix,
  method = "bonferroni",
  correctionLevel = "global",
  filter = TRUE,
  keepSums = FALSE,
  pThreshold = 0.01,
  logThreshold = 1
)
}
\arguments{
\item{dmrMatrix}{Results from testDMR
Should contain a data.table of summed observations per test group and genomic region with unadjusted p values and logFC for each condition.}

\item{method}{Method for p value adjustment. Bonferroni is the default. See ?p.adjust for all options.}

\item{correctionLevel}{Whether to apply multiple testing correction "withinGroup" or "global".}

\item{filter}{If TRUE, removes insignificant results.}

\item{keepSums}{If TRUE, does not remove summed c and t observations per group from the resulting data.table.}

\item{pThreshold}{Maxmimum adjusted p value to allow if filter = TRUE.}

\item{logThreshold}{Minimum absolute value of the log2FC to allow if filter = TRUE.}
}
\value{
Returns an expanded data table with results from testDMR.
}
\description{
Expand and filter the testDMR output
}
\examples{
\dontrun{
  results <- filterDMR(dmrMatrix = dmrs, method = "BH", filter = TRUE, keepSums = FALSE, pThreshold = 0.05, logThreshold = 2)
}
}
