% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{findClusterMarkers}
\alias{findClusterMarkers}
\title{Find cluster markers}
\usage{
findClusterMarkers(obj, matrix, genes)
}
\arguments{
\item{obj}{The amethyst object to analyze. Clustering should be done prior to this step.}

\item{matrix}{Name of the pre-computed methylation values stored genomeMatrices slot for which to calculate variably methylated features}
}
\value{
A data.frame containing the p.value of methylation levels over each gene in each cluster relative to all other clusters
}
\description{
Find variably methylated features across clusters
}
\examples{
clusterMarkers <- findClusterMarkers(obj = obj, matrix = "gene_ch")
}
