% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{findIndex}
\alias{findIndex}
\title{Find index}
\usage{
findIndex(curr_gene, ranges, h5)
}
\arguments{
\item{curr_gene}{Gene name}

\item{ranges}{Ranges data.table}

\item{h5}{A h5 file data.table}
}
\value{
An entry for the index list
}
\description{
Worker function to find a single index used in the indexing function
}
