% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{findRanges}
\alias{findRanges}
\title{findRanges}
\usage{
findRanges(gtf, promoter, subset = NULL)
}
\arguments{
\item{gtf}{Genome annotation file with chromosome, start, and end position
information for genes of interest. See the "makeRef" function.}

\item{promoter}{Bool indicating if the requested type is for 'promoters';
used internally as a flag.}
}
\value{
A data.table of ranges
}
\description{
Worker function to find the ranges used in the indexing function
}
