% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getGeneCoords}
\alias{getGeneCoords}
\title{getGeneCoords}
\usage{
getGeneCoords(ref, gene)
}
\arguments{
\item{ref}{Genome annotation file}

\item{gene}{Name of gene for which to fetch coordinates}
}
\value{
Returns the appended chromosome, start, and end locations for a given gene; e.g. "chr2_199269505_199471266"
}
\description{
Helper function to fetch the appended chromosome, start, and end locations for a given gene
}
\examples{
\dontrun{
 getGeneCoords(ref = obj@ref, gene = "SATB2")
}
}
