% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getGeneM}
\alias{getGeneM}
\title{getGeneM}
\usage{
getGeneM(obj, gene, type, threads = 1, cells = NULL)
}
\arguments{
\item{obj}{The object containing the path to the hdf5 file}

\item{gene}{Which gene to fetch hdf5 indexes}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG.}

\item{cells}{Optional parameter to retrieve indexes for a subset of cells}
}
\value{
Returns a data frame listing the start and count for each barcode with reads covering the requested gene
}
\description{
Helper function to retrieve the relevant portion of the hdf5 file for a given gene
}
\examples{
getGeneM(obj = obj, gene = "RBFOX3", type = "CH", cells = c("ATTGAGGATACCAATTCCATCACGAGCG", "ATTGAGGATAACGCTTCGTCCGCCGATC"))
}
