% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{getRangesFromGenes}
\alias{getRangesFromGenes}
\title{getRangesFromGenes}
\usage{
getRangesFromGenes(ref, chr_col, genes, trackOverhang = 5000)
}
\arguments{
\item{ref}{data.table version of obj@ref.}

\item{chr_col}{Name of chromosome column in ref ("chr" or "seqid").}

\item{genes}{Character vector of gene symbols.}

\item{trackOverhang}{bp to extend on each side of gene coordinates.}
}
\value{
data.table with columns: gene_name, chrom, range_start, range_end.
}
\description{
Convert a set of gene names into genomic ranges using obj@ref.
}
