% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{getRangesFromRegions}
\alias{getRangesFromRegions}
\title{getRangesFromRegions}
\usage{
getRangesFromRegions(regions)
}
\arguments{
\item{regions}{Character vector of "chr_start_end".}
}
\value{
data.table with columns: region_id, chrom, range_start, range_end.
}
\description{
Parse region strings "chr_start_end" into genomic ranges.
}
