% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{getWindowIdxForRanges}
\alias{getWindowIdxForRanges}
\title{getWindowIdxForRanges}
\usage{
getWindowIdxForRanges(idx_dt, ranges_dt)
}
\arguments{
\item{idx_dt}{data.table from buildMetacellH5Index().}

\item{ranges_dt}{data.table with columns: chrom, range_start, range_end.}
}
\value{
Integer vector of unique window indices.
}
\description{
Determine which window indices from an H5 index overlap requested ranges.
}
