% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{heatMap}
\alias{heatMap}
\title{heatMap}
\usage{
heatMap(
  obj,
  genes = NULL,
  regions = NULL,
  matrix,
  colors = NULL,
  trackOverhang = 5000,
  arrowOverhang = 3000,
  nrow = max(length(genes), length(regions)),
  legend = TRUE,
  removeNA = TRUE,
  width = 500
)
}
\arguments{
\item{obj}{Object containing the matrix to plot}

\item{genes}{Gene list to plot}

\item{regions}{Optional. Instead of genes, regions can be plotted, e.g. "chr1_1000_2000".
Any genes or exons that fall within the region will be plotted below.}

\item{matrix}{Matrix contained in the genomeMatrices slot. Note: unlike other functions, designed for makeSlidingWindows output}

\item{colors}{Optional list of colors to include}

\item{trackOverhang}{Number of base pairs to extend beyond the gene}

\item{arrowOverhang}{Number of base pairs the track arrow should extend beyond the gene}

\item{nrow}{When multiple genes or regions are plotted, number of rows the output is divided across.}

\item{legend}{Boolean indicating whether to plot the color legend}

\item{removeNA}{Boolean indicating whether to remove NA values as opposed to plotting them grey}

\item{width}{Width of the tiles, which should correspond to the step size calculated with calcSmoothedWindows.}
}
\value{
A ggplot geom_tile object with colors indicating \% methylation over 500 bp windows and the gene of interest beneath
}
\description{
Plot methylation levels aggregated across smoothed genomic windows over a given gene body or genomic region
}
\examples{
heatMap(obj, matrix = "cluster_cg_500_slidingwindows", regions = "chr11_47369500_47420000")
}
