% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{histGeneM}
\alias{histGeneM}
\title{histGeneM}
\usage{
histGeneM(
  obj,
  type,
  groupBy = NULL,
  genes,
  bins = 100,
  colors = pal,
  smooth = TRUE,
  track = TRUE,
  promoter = FALSE,
  track.color = "cornflowerblue",
  alpha = 0.2,
  nrow = length(genes),
  legend = F
)
}
\arguments{
\item{obj}{Amethyst object to plot}

\item{type}{Type of methylation to plot - either "CG", "CG", or "both"}

\item{groupBy}{A categorical variable contained in the metadata to facet the plot by}

\item{genes}{Which genes to plot methylation over}

\item{bins}{Number of bins to divide the histogram by}

\item{colors}{Optional color palette to plot}

\item{smooth}{Boolean indicating whether to plot a smoothed average line of histogram values}

\item{track}{Boolean indicating whether to plot a gene track below the histogram}

\item{promoter}{Boolean indicating whether to plot promoter CG levels (appears as a red block)}

\item{track.color}{Option to change the color in which the gene track is plotted}

\item{alpha}{Opacity of the histogram}

\item{nrow}{Number of rows to divide the output across}

\item{legend}{Boolean indicating whether to plot the groupBy variable legend}
}
\value{
Returns a ggplot histogram showing methylation levels over a gene body
}
\description{
Plot methylation levels over a gene body in histogram format
}
\examples{
histGeneM(obj = obj, type = "both", groupBy = cluster_id, genes = "SATB2")
}
