% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{histograM}
\alias{histograM}
\title{histograM}
\usage{
histograM(
  obj,
  matrix,
  genes = NULL,
  regions = NULL,
  colors = NULL,
  trackOverhang = 5000,
  arrowOverhang = 3000,
  legend = TRUE,
  removeNA = TRUE,
  trackScale = 0.07,
  arrowScale = NULL,
  colorMax = NULL,
  order = NULL,
  extraTracks = NULL,
  remove = NULL,
  baseline = "mean",
  nrow = max(length(genes), length(regions)),
  orientation = "rows"
)
}
\arguments{
\item{obj}{Object containing the matrix to plot}

\item{matrix}{Matrix contained in the genomeMatrices slot. Note: unlike other functions, designed for makeSlidingWindows output}

\item{genes}{Gene list to plot}

\item{regions}{Optional. Instead of genes, regions can be plotted, e.g. "chr1_1000_2000".
Any genes or exons that fall within the region will be plotted below.}

\item{colors}{Optional list of colors to include}

\item{trackOverhang}{Number of base pairs to extend beyond the gene}

\item{arrowOverhang}{Number of base pairs the track arrow should extend beyond the gene}

\item{legend}{Boolean indicating whether to plot the color legend}

\item{removeNA}{Boolean indicating whether to remove NA values as opposed to plotting them grey}

\item{trackScale}{Enables adjustment of gene track height}

\item{arrowScale}{Enables adjustment of gene directional arrow size}

\item{colorMax}{Set upper bound on color scale. Everything exceeding this threshold will be plotted at the max value}

\item{order}{Arrange the order of groups}

\item{extraTracks}{Optional; include additional tracks (like regulatory elements) below gene tracks.
Input should be a named list of coordinate files containing chr, start, and end columns.}

\item{remove}{Option to remove genes containing specific patterns from the plot, e.g. "^ENS|^LINC".
Helpful when plotting very large regions.}

\item{baseline}{Option for histogram to begin at 0 or at the "mean" methylation values for that group.}

\item{nrow}{When multiple genes or regions are plotted, number of rows the output is divided across.}

\item{orientation}{Defines whether groups are plotted in "rows" or "cols".}
}
\value{
A ggplot object with color and bar height indicating \% methylation over windows.
Genes and any imported tracks (optional) are plotted beneath.
}
\description{
Plot methylation levels aggregated across smoothed genomic windows over a given gene body or genomic region
}
\examples{
histograM(brain, matrix = "ch_type_tracks", genes = "SATB2")
histograM(brain, matrix = "ch_type_tracks", regions = c("chr2_170813213_170861151", "chr2_199269505_199471266"), orientation = "cols", trackOverhang = 1000000, remove = "ENS|LINC")
}
