% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R, R/helper_dt.R
\name{impute}
\alias{impute}
\title{Impute}
\usage{
impute(obj, matrix, npca = 10)

impute(obj, matrix, npca = 10)
}
\arguments{
\item{obj}{Amethyst object to perform imputation on}

\item{matrix}{Name of the matrix contained in the genomeMatrices slot to perform imputation on}

\item{npca}{Number of principle components to use}
}
\value{
Adds a new data frame to the genomeMatrices slot with imputed values

Adds a new data frame to the genomeMatrices slot with imputed values
}
\description{
Impute values with the Rmagic package https://github.com/KrishnaswamyLab/MAGIC

Impute values with the Rmagic package https://github.com/KrishnaswamyLab/MAGIC
}
\examples{
obj <- impute(obj, matrix = "gene_ch")
obj <- impute(obj, matrix = "gene_ch")
}
