% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{indexAll}
\alias{indexAll}
\title{indexAll}
\usage{
indexAll(
  hdf5,
  gtf,
  subset = markerGenes,
  threads = 10,
  save = NULL,
  method = "foreach",
  choice = "all"
)
}
\arguments{
\item{hdf5}{Path to the hdf5 file containing base-level read information organized by methylation type and barcode.}

\item{subset}{List of genes to index. Optional but recommended.See the "fetchMarkers" function.}

\item{threads}{Optional threads parameter to use if parallelizing. Highly recommended as even small datasets will take a long time to index.}

\item{save}{Optional path to include for saving the intermediate index output as an RData file.}
}
\value{
Returns a list of coordinates in the hdf5 file corresponding to each gene. Organized by methylation type, gene, and cell.
}
\description{
Index gene body CG methylation, gene body CH methylation, and mCG status of promoters.
Promoters are defined by the first position of the gene body (strand-aware) +/- 1500 bp.
}
\examples{
index <- indexAll(hdf5 = "~/Downloads/test.h5", subset = c("SATB2", "TBR2", "GAD1"), threads = 20, save = "~/Downloads/test_index.RData")
}
