% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{indexChr}
\alias{indexChr}
\title{indexChr}
\usage{
indexChr(obj, type, chrList = NULL, threads = 1)
}
\arguments{
\item{obj}{Amethyst object containing h5 paths}

\item{type}{What type of methylation to retrieve; e.g. CG or CH.}

\item{chrList}{Optional; chromosome whitelist (only use if necessary)}

\item{threads}{Optional; number of threads to use if parallelizing (recommended)}
}
\value{
A list of data frames named by cell barcode. The data frames contain the
coordinates in the hdf5 file corresponding to each chromosome.
}
\description{
If the whole hdf5 file had to be searched for relevant reads
every time gene-specific methylation information was needed, most functions
would take minutes to hours to run. The indexing process catalogs the
coordinates for each specified gene in every cell beforehand so the relevant
subset can quickly be retrieved for downstream processes. Typically, CG and CH
contexts would be indexed. This step can be skipped if using pre-computed
windows with Facet.
}
\examples{
\dontrun{
  obj@index[["chr_cg"]] <- indexChr(obj = obj, type = "CG", threads = 10)
  obj@index[["chr_ch"]] <- indexChr(obj = obj, type = "CG", chrList = c("chr1", "chr2"))
}
}
