% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{indexGenes}
\alias{indexGenes}
\title{indexGenes
If the whole hdf5 file had to be searched for relevant reads every time gene-specific methylation information was needed,
most functions would take minutes to hours to run. The indexing process catalogs the coordinates for each specified gene in
every cell beforehand so the relevant subset can quickly be retrieved for downstream processes. mCG, mCH, and the mCG status
of the promoter must be indexed separately. Providing a gene subset is recommended due to memory and time constraints.}
\usage{
indexGenes(hdf5, gtf = ref, type, threads = 10, subset = markerGenes)
}
\arguments{
\item{hdf5}{Path to the hdf5 file containing base-level read information organized by methylation type and barcode.}

\item{gtf}{Genome annotation file with chromosome, start, and end position information for genes of interest. See the "makeRef" function.}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG.}

\item{threads}{Optional threads parameter to use if parallelizing. Highly recommended as even small datasets will take a long time to index.}

\item{subset}{List of genes to index. Optional but recommended.See the "fetchMarkers" function.}
}
\value{
Returns a list of dataframes named by gene. The dataframes contain the coordinates in the hdf5 file corresponding to the gene's genomic
location for each cell.
}
\description{
indexGenes
If the whole hdf5 file had to be searched for relevant reads every time gene-specific methylation information was needed,
most functions would take minutes to hours to run. The indexing process catalogs the coordinates for each specified gene in
every cell beforehand so the relevant subset can quickly be retrieved for downstream processes. mCG, mCH, and the mCG status
of the promoter must be indexed separately. Providing a gene subset is recommended due to memory and time constraints.
}
\examples{
index[["CH"]] <- indexGenes(hdf5 = "~/Downloads/test.h5", gtf = ref, type = "CH", threads = 10, subset = c("SATB2", "TBR1", "FOXG1"))
}
