% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{loadSmoothedWindows}
\alias{loadSmoothedWindows}
\title{loadSmoothedWindows}
\usage{
loadSmoothedWindows(
  obj,
  type = "CG",
  threads = 1,
  name = "500",
  smoothTrim = 500,
  futureType = "multicore",
  groupBy = "cluster_id",
  nmin = 10,
  chunkSize = 50,
  returnSumMatrix = TRUE,
  returnPctMatrix = TRUE
)
}
\arguments{
\item{obj}{Amethyst object containing the h5paths to the pre-aggregated windows}

\item{type}{Type of methylation - e.g. "CG" or "CH" - to calculate. Typically will be "CG"}

\item{threads}{Optional number of threads for parallelization (multithreading strongly suggested)}

\item{name}{Name of aggregated dataset in h5 file}

\item{smoothTrim}{Width of values to trim for output in the "start" and "end" columns. For example, we typically
calculate a 1500x500bp smoothed sliding window matrix, so 500 is trimmed off either end to produce smoothed unique values}

\item{futureType}{If using parallelization, should R multithread using "multicore" or "multisession"}

\item{groupBy}{Parameter contained in the metadata over which to aggregate observations. Default is "cluster_id"}

\item{nmin}{Minimum number of observations to include window value}

\item{chunkSize}{How many cells to process at a time (for memory management purposes)}

\item{returnSumMatrix}{Whether or not the function should return the matrix of summed c and t observations. Required for testDMR input}

\item{returnPctMatrix}{Whether or not the function should calculate \% methylation over each genomic window. Required for heatMap input}
}
\value{
if returnSumMatrix = TRUE, returns a data.table with the genomic location as chr, start, and end columns,
plus aggregated c and t observations for each groupBy value. If returnPctMatrix = TRUE, returns a data.table with the
genomic location as chr, start, and end columns, plus one column of the mean \% methylation for each window per groupBy value.
If both are true, returns a list containing each result.
}
\description{
Aggregate c and t observations over smoothed genomic windows of a user-defined size.
Optionally calculate the \% methylation for each window.
}
\examples{
\dontrun{
  cluster1kbwindows <- loadSmoothedWindows(obj, type = "CG", threads = 10, name = "500", smoothTrim = 500)
}
}
