% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{loadWindows}
\alias{loadWindows}
\title{loadWindows}
\usage{
loadWindows(
  obj,
  type = "CG",
  metric = "percent",
  threads = 1,
  name,
  futureType = "multicore",
  nmin = 10,
  chunkSize = 50
)
}
\arguments{
\item{obj}{Amethyst object containing the h5paths to the pre-aggregated windows}

\item{type}{Type of methylation - e.g. "CG" or "CH" - to calculate. Typically will be "CG"}

\item{metric}{Calculate either methylation "percent" (out of 100), "score" (normalized metric from -1 to 1 with
0 as the mean), or "ratio" (value / global methylation)}

\item{threads}{Optional number of threads for parallelization}

\item{name}{Name of aggregated dataset in h5 file}

\item{futureType}{If using parallelization, should R multithread using "multicore" or "multisession".}

\item{nmin}{Minimum number of observations to include window value}

\item{chunkSize}{How many cells to process at a time (for memory management purposes)}
}
\value{
Returns a data frame with columns as cells, rows as genomic windows, and values as aggregated methylation levels
}
\description{
Load pre-computed c and t observations aggregated with Facet
}
\examples{
\dontrun{
  obj@genomeMatrices[["cg_100k_score_facet"]] <- loadWindows(obj, name = "100000", nmin = 2, metric = "score")
}
}
