% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{makeAmethystPalette}
\alias{makeAmethystPalette}
\title{Make amethyst palette}
\usage{
makeAmethystPalette()
}
\value{
A vector of 13 colours
}
\description{
Returns a predefined palette frequently used in the \code{amethyst} package
}
\examples{
pal <- makeAmethystPalette()
}
