% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makeClusterTracks}
\alias{makeClusterTracks}
\title{makeClusterTracks}
\usage{
makeClusterTracks(obj, type, threads = 1, bed, index)
}
\arguments{
\item{obj}{Amethyst object containing the path(s) to calculate.}

\item{type}{Type of methylation - e.g. "CG" or "CH" - to calculate. Typically will be "CG."}

\item{threads}{Optional number of threads for parallelization. Basically required for this step.}

\item{bed}{Bed file containing genome in 500bp windows. Options are "hg38", "mm10", or path to bed file.}

\item{index}{Name of chr index in the index slot.
This reduces memory constraints by processing one chromosome at a time.}
}
\value{
Returns a data.table with the genomic location as chr, start, and end columns, plus one column of the mean \% methylation per cluster
}
\description{
Calculate mean percent methylation over 1500 base pair genomic windows shifted every 500 bp and aggregate by cluster.
Intended to generate input for the tileGeneM function.
}
\examples{
cluster_500bp_cg_tracks <- makeClusterTracks(obj = obj, type = "CG", threads = 20, bed = "hg38")
}
