% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.r
\name{makeDoubletObject}
\alias{makeDoubletObject}
\title{makeDoubletObject}
\usage{
makeDoubletObject(
  obj,
  simFraction = 0.25,
  threads = 10,
  genomeMatrices = c("cg_100k_score", "ch_100k_pct")
)
}
\arguments{
\item{obj}{An amethyst object containing the original cell data.}

\item{simFraction}{A numeric value specifying the fraction of cells to simulate as doublets. Default is 0.25.}

\item{threads}{An integer specifying the number of threads to use for parallel processing. Default is 10.}

\item{genomeMatrices}{A character vector specifying the names of the genome matrices to use. Default is c('cg_100k_score','ch_100k_pct').}
}
\value{
An amethyst object containing both true cells and artificial doublets.
}
\description{
This function makes the new object with true cells as well as artificial doublets
}
\examples{
\dontrun{
dbobj <- makeDoubletObject(obj, simFraction=0.25, threads = 10, genomeMatrices=c("cg_100kb_score", "ch_100k_pct"))
}
}
