% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{makePalette}
\alias{makePalette}
\title{makePalette}
\usage{
makePalette(option, n, sample = T)
}
\arguments{
\item{option}{Which palette to choose out of options 1-30. Use testPalette to see the colors in each}

\item{n}{How many colors are needed in the palette}
}
\value{
Returns a character vector of n hex codes
}
\description{
Generate a color palette of any length from a pre-defined set of hex codes
}
\examples{
\dontrun{
  pal <- makePalette(option = 1, n = 20)
}
}
