% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makeSlidingWindows}
\alias{makeSlidingWindows}
\title{makeSlidingWindows}
\usage{
makeSlidingWindows(obj, type, threads = 1)
}
\arguments{
\item{obj}{Amethyst object containing the path to the h5 file over which to calculate 500 bp windows}

\item{type}{Type of methylation to calculate; e.g. "CG" or "CH"}

\item{threads}{Enables multithreading}
}
\value{
Returns a data frame with \% methylation over 1500 base pair genomic windows shifted every 500 bp aggregated by cluster
}
\description{
Calculate mean percent methylation over 1500 base pair genomic windows shifted every 500 bp and aggregate by cluster.
Intended to generate input for the tileGeneM function.
}
\examples{
cluster_cg_500bp_windows <- makeSlidingWindows(obj = obj, type = "CG", threads = 20)
}
