% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{markerHeatmap}
\alias{markerHeatmap}
\title{markerHeatmap
Plot the methylation levels of sample groups across many genes at once}
\usage{
markerHeatmap(
  obj,
  level = "CellClass",
  impute = TRUE,
  rownames = FALSE,
  type = "CH"
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
markerHeatmap
Plot the methylation levels of sample groups across many genes at once
}
