% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{metacellMeanReduction}
\alias{metacellMeanReduction}
\title{metacellMeanReduction}
\usage{
metacellMeanReduction(
  metacell_obj,
  original_obj,
  original_reduction,
  metacell_reduction
)
}
\arguments{
\item{metacell_obj}{Metacell Amethyst object (from makeMetacells()).}

\item{original_obj}{Original single-cell Amethyst object (with reduction to average).}

\item{original_reduction}{Name of reduction in original_obj@reductions to average.}

\item{metacell_reduction}{Name to give the new reduction in metacell_obj@reductions.}
}
\value{
The metacell_obj with a new reduction added.
}
\description{
Recompute a reduction on the metacell object by averaging the original reduction
across member cells of each metacell.
}
