% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{noAxes}
\alias{noAxes}
\title{noAxes}
\usage{
noAxes(..., keep.text = FALSE, keep.ticks = FALSE)
}
\description{
Add to a ggplot object to remove axes
Developed by the Satija lab https://github.com/satijalab/seurat/blob/master/R/visualization.R
}
