% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_dt.R
\name{roundAny}
\alias{roundAny}
\title{Round any}
\usage{
roundAny(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric vector to be rounded}

\item{accuracy}{number to round to}

\item{f}{rounding function}
}
\value{
x with numbers rounded to \code{accuracy} number of significant figures
as defined by \code{f}
}
\description{
Local implementation of the plyr function \code{round_any}.
}
\examples{
roundAny(111.1111, 1)
roundAny(111.1111, 10)
roundAny(111.1111, 100)
roundAny(111.1111, 0.1)
roundAny(111.1111, 1, floor)
roundAny(111.1111, 0.1, floor)
roundAny(111.1111, 1, ceiling)
roundAny(111.1111, 100, ceiling)
roundAny(111.1111, 0.1, ceiling)
}
