% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{runIrlba}
\alias{runIrlba}
\title{runIrlba}
\usage{
runIrlba(obj, genomeMatrices, dims)
}
\arguments{
\item{obj}{Object for which to run irlba}

\item{genomeMatrices}{list of matrices in the genomeMatrices slot to use for irlba}

\item{dims}{list of how many dimensions to output for each matrix}
}
\value{
Returns a matrix of appended irlba dimensions as columns and cells as rows
}
\description{
Perform dimensionality reduction based on methylation levels over a matrix stored in the @genomeMatrices slot
}
\examples{
obj <- runIrlba(obj, genomeMatrices = c("ch_2M_pct", "ch_2M_score", "cg_2M_score"), dims = c(10, 10, 10))
}
