% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{sampleComp}
\alias{sampleComp}
\title{sampleComp}
\usage{
sampleComp(obj, groupBy, colorBy, colors = NULL)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{groupBy}{What column in the metadata to group cells by. This will be on the x axis.}

\item{colorBy}{What column in the metadata to color the bar chart by.}

\item{colors}{Optional color palette}
}
\value{
A bar graph showing the distribution of categorical metadata in user-defined groups of cells
}
\description{
Visualize the distribution of a metadata variable across groups of cells
}
\examples{
\dontrun{
  sampleComp(obj = obj, groupBy = "sample", colorBy = "cluster_id")
  sampleComp(obj = obj, groupBy = "sample", colorBy = "cluster_id") + facet_wrap(vars(batch))
}
}
