% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subsetObject}
\alias{subsetObject}
\title{subsetObject}
\usage{
subsetObject(obj, cells)
}
\arguments{
\item{obj}{Amethyst object to subset}

\item{cells}{Character vector of barcodes to include in the subset}
}
\value{
Returns a new amethyst object with all slots subsetted
}
\description{
Subset all slots of an amethyst object to only include information from specified barcodes
}
\examples{
subset <- subsetObject(obj = obj, cells = c("ATTGAGGATAACGCTTCGTCCGCCGATC", "ATTGAGGATAACGCTTCGTCTAAGGTCA"))
}
