% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{tileGeneM}
\alias{tileGeneM}
\title{tileGeneM}
\usage{
tileGeneM(
  obj,
  genes,
  matrix,
  colors = NULL,
  trackOverhang = 5000,
  arrowOverhang = 3000,
  nrow = length(genes),
  legend = TRUE
)
}
\arguments{
\item{obj}{Object containing the matrix to plot}

\item{genes}{Gene list to plot}

\item{matrix}{Matrix contained in the genomeMatrices slot. Note: unlike other functions, designed for makeSlidingWindows output}

\item{colors}{Optional list of colors to include}

\item{trackOverhang}{Number of base pairs to extend beyond the gene}

\item{arrowOverhang}{Number of base pairs the track arrow should extend beyond the gene}
}
\value{
A ggplot geom_tile object with colors indicating \% methylation over 500 bp windows and the gene of interest beneath
}
\description{
Plot CG methylation levels over a gene body in 500bp smoothed windows
}
\examples{
tileGeneM(obj, gene = "SYT7", matrix = "cluster_cg_500_slidingwindows")
}
