% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additions.R
\name{transferLabelsNN}
\alias{transferLabelsNN}
\title{transferLabelsNN}
\usage{
transferLabelsNN(
  obj,
  labels = "labels",
  transferred = "transferred",
  nnn = 5,
  reduction = "irlba"
)
}
\arguments{
\item{obj}{Amethyst object on which to perform label transfer. The object must contain metadata with the specified labels and an \code{irlba} slot.}

\item{labels}{A character string specifying the metadata column in the Amethyst object that contains the labels to be transferred. Defaults to 'labels'.}

\item{transferred}{A character string specifying the name of the new column in the metadata slot where the transferred labels will be stored. Defaults to 'transferred'.}

\item{nnn}{An integer specifying the maximum number of nearest neighbors to search. Defaults to 5.}
}
\value{
The Amethyst object with a new column in the metadata containing the transferred labels.
}
\description{
Transfer metadata labels based on nearest neighbor analysis.
}
