% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{umapGeneM}
\alias{umapGeneM}
\title{umapGeneM}
\usage{
umapGeneM(
  obj,
  genes,
  type = NULL,
  impute = FALSE,
  matrix = NULL,
  blend = FALSE,
  colors = NULL,
  nrow = round(sqrt(length(genes)))
)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{genes}{List of genes to plot \% methylation in UMAP space}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG.}

\item{impute}{Logical indicating whether to impute values using Rmagic}

\item{matrix}{Pre-computed matrix to use in the genomeMatrices slot, if available. The type parameter is not needed if matrix is definied.}

\item{blend}{Logical indicating whether to blend methylation levels of two or three genes}

\item{nrow}{Number of rows to distribute graphs over}
}
\value{
Returns a plot of methylation over given genomic region in relation to UMAP
}
\description{
Plot cumulative \% methylation over a gene body on umap
}
\examples{
umapGeneM(obj, genes = c("GAD1", "SATB2"), type = "CH")
umapGeneM(obj, genes = c("GAD1", "SATB2"), type = "CH", blend = TRUE, nrow = 1)
umapGeneM(both, genes = c("Cux1", "Lingo1"), matrix = "genemeans_cac_imputed", blend = T)

}
