% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{violinM}
\alias{violinM}
\title{violinM}
\usage{
violinM(
  obj,
  genes,
  matrix,
  groupBy,
  colors = NULL,
  pointSize = 0.2,
  nrow = round(sqrt(length(genes)))
)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{genes}{Genes to plot}

\item{matrix}{Optional name of a pre-calculated matrix of values contained in the genomeMatrices slot to use}

\item{groupBy}{Parameter to group by on the x axis}

\item{colors}{Optional color palette}

\item{nrow}{Number of rows to plot if visualizing many genes}

\item{type}{Type of methylation - e.g. "CH" or "CG" - to calculate if matrix is not provided}
}
\value{
Returns ggplot2 geom_violin plots of methylation levels faceted by the requested genes
}
\description{
Generates a violin plot of percent methylation over a gene body. Best for non-CpG methylation.
}
\examples{
violinM(obj, genes = c("GRM8"), matrix = "gene_ch", groupBy = "cluster_id")
}
