% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{vlnGeneM}
\alias{vlnGeneM}
\title{vlnGeneM}
\usage{
vlnGeneM(
  obj,
  genes,
  type = NULL,
  metric = "percent",
  matrix = NULL,
  groupBy,
  colors = NULL,
  nrow = round(sqrt(length(genes)))
)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{genes}{Genes to plot}

\item{type}{Type of methylation - e.g. "CH" or "CG" - to calculate if matrix is not provided}

\item{matrix}{Optional name of a pre-calculated matrix of values contained in the genomeMatrices slot to use}

\item{groupBy}{Parameter to group by on the x axis}

\item{colors}{Optional color palette}

\item{nrow}{Number of rows to plot if visualizing many genes}
}
\value{
Returns ggplot2 geom_violin plots of methylation levels faceted by the requested genes
}
\description{
Generates a violin plot of percent methylation over a gene body
}
\examples{
vlnGeneM(obj, genes = c("SATB2", "GAD1"), matrix = "test_gene_ch", groupBy = "cluster_id")
}
