% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addAnnot}
\alias{addAnnot}
\title{addAnnot}
\usage{
addAnnot(obj, file)
}
\arguments{
\item{obj}{Amethyst object to add metadata to}

\item{file}{Path to .annot file}
}
\value{
Updates obj to contain the new metadata concatenated to the old metadata
}
\description{
Add information containined in the .annot file to the Amethyst object
}
\examples{
obj <- addMetadata(obj = obj, metadata = annotations)
}
