% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R, R/helper_dt.R
\name{addCellInfo}
\alias{addCellInfo}
\title{Add cell info}
\usage{
addCellInfo(obj, file)

addCellInfo(obj, file)
}
\arguments{
\item{obj}{Amethyst object to add cell info metadata}

\item{file}{Path of the cellInfo.txt file}
}
\value{
Returns the same Amethyst object but with cellInfo added

Returns the same Amethyst object but with cellInfo added
}
\description{
Add the information contained in the cellInfo file outputs to the obj@metadata

Add the information contained in the cellInfo file outputs to the obj@metadata
}
\examples{
obj <- addCellInfo(obj, file = "~/Downloads/cellInfo.txt")
obj <- addCellInfo(obj, file = "~/Downloads/cellInfo.txt")
}
