% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{aggregateMatrix}
\alias{aggregateMatrix}
\title{aggregateMatrix}
\usage{
aggregateMatrix(obj, matrix, groupBy, name)
}
\arguments{
\item{obj}{Amethyst pbject containing the matrix to be aggregated}

\item{matrix}{Name of matrix contained in the genomeMatrices slot to aggregate}

\item{groupBy}{Parameter in the metadata to groupBy}

\item{name}{Name of output matrix to store}
}
\description{
aggregateMatrix
}
