% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R, R/helper_dt.R
\name{aggregateMatrix}
\alias{aggregateMatrix}
\title{aggregateMatrix}
\usage{
aggregateMatrix(obj, matrix, groupBy, name)

aggregateMatrix(obj, matrix, groupBy, name)
}
\arguments{
\item{obj}{Amethyst object containing the matrix to be aggregated}

\item{matrix}{Name of matrix contained in the genomeMatrices slot to aggregate}

\item{groupBy}{Parameter in the metadata to groupBy}

\item{name}{Name of output matrix to store}
}
\value{
Returns a matrix containing the mean value per group

Amethyst object with aggregated genome matrices
}
\description{
Aggregate genomic window matrices by a categorical variable contained in the metadata
}
\examples{
obj <- aggregateMatrix(obj, matrix = "ch_100k_pct", groupBy = cluster_id, name = "cluster_ch_100k_pct")
\dontrun{
obj <- aggregateMatrix(obj, matrix = "ch_100k_pct", groupBy = cluster_id, name = "cluster_ch_100k_pct")
}
}
