% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{aggregateMatrix}
\alias{aggregateMatrix}
\title{aggregateMatrix}
\usage{
aggregateMatrix(obj, matrix, groupBy, minCells = 5, minValues = 5)
}
\arguments{
\item{obj}{Amethyst object containing the matrix to be aggregated}

\item{matrix}{Name of matrix contained in the genomeMatrices slot to aggregate}

\item{groupBy}{Parameter in the metadata to aggregate over}

\item{minCells}{Minimum number of cells in the group to return an aggregated output}

\item{minValues}{Minimum number of non-NA values within a group to return an aggregated value}
}
\value{
Returns a matrix containing the mean value per group
}
\description{
Aggregate genomic window matrices by a categorical variable contained in the metadata
}
\examples{
\dontrun{
  obj@genomeMatrices[["ch_100k_pct_cluster"]] <- aggregateMatrix(obj, matrix = "ch_100k_pct", groupBy = "cluster_id", nmin = 10)
}
}
