% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{collapseDMR}
\alias{collapseDMR}
\title{collapseDMR}
\usage{
collapseDMR(dmrMatrix, max_dist = 0, min_length = 1000, collapse = TRUE)
}
\arguments{
\item{dmrMatrix}{Table of differentially methylated regions identified with testDMR and filterDMR}

\item{max_dist}{Maximum allowable overlap between DMRs to be considered adjacent}

\item{min_length}{Minimum length of collapsed DMR window to include in the output}
}
\value{
Returns a data.table of collapsed DMRs
}
\description{
Identify adjacent DMRs
}
\examples{
results <- collapseDMR(dmrs, max_dist = 2000, min_length = 5000, collapse = T)
}
