% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{collapseDMR}
\alias{collapseDMR}
\title{collapseDMR}
\usage{
collapseDMR(
  obj,
  dmrMatrix,
  maxDist = 0,
  minLength = 1000,
  reduce = TRUE,
  mergeDirections = FALSE,
  annotate = TRUE
)
}
\arguments{
\item{dmrMatrix}{Table of differentially methylated regions identified with testDMR and filterDMR}

\item{maxDist}{Maximum allowable overlap between DMRs to be considered adjacent}

\item{minLength}{Minimum length of collapsed DMR window to include in the output}
}
\value{
Returns a data.table of collapsed DMRs
}
\description{
Identify adjacent DMRs
}
\examples{
results <- collapseDMR(dmrs, maxDist = 2000, minLength = 2000, reduce = T, annotate = T)
}
