% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convertObject}
\alias{convertObject}
\title{convertObject}
\usage{
convertObject(obj)
}
\arguments{
\item{obj}{Old amethyst object to convert}
}
\value{
Returns an updated amethyst object with slots:
h5paths, genomeMatrices, tracks, reductions, index, metadata, ref, results
}
\description{
Shortcut for converting old amethyst object format (v0.0.0.9000) to new (v1.0.0)
Adds "tracks" and "results" slots, converts obj@ref to a data.table, moves tracks and umap/tsne
coordinates to correct locations if the names contain "tracks" and "umap" or "tsne", respectively
}
\examples{
\dontrun{
 new_obj <- convertObject(old_obj)
}
}
