% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{createObject}
\alias{createObject}
\title{createObject}
\usage{
createObject(h5path, index = NULL, ref = NULL, metadata = NULL)
}
\arguments{
\item{h5path}{Path to the hdf5 file containing base-level read information organized by methylation type and barcode}

\item{index}{Corresponding gene coordinates in the hdf5 file}

\item{ref}{Genome annotation file with chromosome, start, and end position information for genes of interest. See the "makeRef" function.}

\item{metadata}{Optional cell metadata. If included, make sure row names are cell IDs}
}
\value{
Returns a single object of class amethyst with h5path, index, metadata, ref, reduction, and irlba slots.
}
\description{
Create object of class amethyst
}
\examples{
obj <- createObject(h5path = "~/Downloads/test.h5")
}
