% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{dimFeature}
\alias{dimFeature}
\title{dimFeature}
\usage{
dimFeature(
  obj,
  colorBy = NULL,
  colors = NULL,
  labelBy = NULL,
  pointSize = 0.5,
  reduction = "umap"
)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{colorBy}{The metadata feature to plot over UMAP / TSNE coordinates}

\item{colors}{Optional color scale}

\item{pointSize}{Optional; adjust point size}

\item{reduction}{Name of reduction slot containing UMAP / TSNE dimensionality reduction coordinates}

\item{label}{Optional; text label overlay}
}
\value{
Returns a ggplot graph plotting xy coordinates of cells colored according to the specified feature
}
\description{
Plot any feature in the metadata over UMAP / TSNE coordinates
}
\examples{
\dontrun{
  dimFeature(obj = obj, colorBy = "cluster_id", reduction = "umap")
}

}
