% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{dimM}
\alias{dimM}
\title{dimM}
\usage{
dimM(
  obj,
  genes,
  matrix,
  blend = FALSE,
  colors = NULL,
  squish = NULL,
  reduction = "umap",
  pointSize = 0.1,
  nrow = round(sqrt(length(genes)))
)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{genes}{List of genes to plot \% methylation in UMAP or TSNE space}

\item{matrix}{Pre-computed matrix to use in the genomeMatrices slot, if available. The type parameter is not needed if matrix is definied.}

\item{blend}{Logical indicating whether to blend methylation levels of two or three genes}

\item{colors}{Optional color gradient to include}

\item{nrow}{Number of rows to distribute graphs over}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG.}

\item{metric}{If matrix is not specified, what methylation parameter to calculate, i.e. "percent", "score", or "ratio"}
}
\value{
Returns a plot of methylation over given genomic region using UMAP or TSNE coordinates
}
\description{
Plot cumulative \% methylation over a gene body on UMAP or TNSE dimensionality reductions
}
\examples{
dimM(obj, genes = c("GRM8", "SATB1"), matrix = "gene_ch", pointSize = 0.8)
}
