% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{dotM}
\alias{dotM}
\title{dotM}
\usage{
dotM(obj, genes, groupBy, matrix)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{genes}{List of genes to plot methylation on the x axis}

\item{groupBy}{A categorical variable contained in the metadata to group cells by on the y axis}

\item{matrix}{Name of a pre-calculated methylation levels over genes of interest contained in the genomeMatrices slot}
}
\value{
Returns a ggplot object displaying average methylation values for each gene by the grouping variable
}
\description{
Generates a dot plot of percent methylation over a gene body
}
\examples{
dotM(obj,  genes = c("GRM8", "SATB2", "LINGO1"), matrix = "gene_ch", groupBy = "cluster_id")
}
