% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{findClusterMarkers}
\alias{findClusterMarkers}
\title{Find cluster markers}
\usage{
findClusterMarkers(
  obj,
  impute = FALSE,
  genes = names(obj@index[["CH"]]),
  nmin = NULL,
  type = NULL,
  matrix = NULL
)
}
\arguments{
\item{obj}{The amethyst object to analyze. Clustering should be done prior to this step.}

\item{impute}{Optional parameter to address coverage issues. If TRUE, Rmagic will be used to impute values based on available data.}

\item{genes}{Which genes to test. Genes must be indexed.}

\item{nmin}{Optional minimum threshold}

\item{type}{}
}
\value{
A data.frame of ...
}
\description{
Find variably methylated features across clusters
}
