% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{heatMap}
\alias{heatMap}
\title{heatMap}
\usage{
heatMap(
  obj,
  track = NULL,
  genes = NULL,
  regions = NULL,
  colors = NULL,
  trackOverhang = 5000,
  arrowOverhang = 3000,
  legend = TRUE,
  removeNA = TRUE,
  trackScale = 0.07,
  arrowScale = NULL,
  colorMax = NULL,
  order = NULL,
  extraTracks = NULL,
  extraTrackColors = NULL,
  remove = NULL,
  subset = NULL,
  nrow = max(length(genes), length(regions))
)
}
\arguments{
\item{obj}{Object containing the matrix to plot}

\item{track}{Name of matrix containing aggregated methylation values calculated over small genomic windows}

\item{genes}{Gene list to plot}

\item{regions}{Optional. Instead of genes, regions can be plotted, e.g. "chr1_1000_2000".
Any genes or exons that fall within the region will be plotted below.}

\item{colors}{Optional; specify color palette}

\item{trackOverhang}{Number of base pairs to extend beyond the gene}

\item{arrowOverhang}{Number of base pairs the track arrow should extend beyond the gene}

\item{legend}{Boolean indicating whether to plot the color legend}

\item{removeNA}{Boolean indicating whether to remove NA values as opposed to plotting them grey}

\item{trackScale}{Enables adjustment of gene track height}

\item{arrowScale}{Enables adjustment of gene directional arrow size}

\item{colorMax}{Set upper bound on color scale. Everything exceeding this threshold will be plotted at the max value}

\item{order}{Arrange the order of groups}

\item{extraTracks}{Optional; include additional tracks (like regulatory elements) below gene tracks.
Input should be a named list of coordinate files containing chr, start, and end columns.}

\item{extraTrackColors}{Optional vector of hex code colors to use when plotting extraTracks}

\item{remove}{Option to remove genes containing specific patterns from the plot, e.g. "^ENS|^LINC".
Helpful when plotting very large regions.}

\item{subset}{Option to subset tracks to groups containing specific text patterns, e.g. "^Pvalb|^Sst".}

\item{nrow}{When multiple genes or regions are plotted, number of rows the output is divided across.}
}
\value{
A ggplot geom_tile object with colors indicating \% methylation over tiled windows and the gene of interest beneath
}
\description{
Plot methylation levels aggregated across a given gene body or genomic region
}
\examples{
\dontrun{
 heatMap(obj, regions = "chr1_153390032_153391073", remove = "ENS", track = "cg_cluster_tracks", extraTracks = ccre_tracks)
 heatMap(obj, genes = c("SATB2", "GAD1"), track = "cg_type_tracks", extraTracks = brain_dmrs)
}
}
