% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{heatMapGenome}
\alias{heatMapGenome}
\title{heatMapGenome}
\usage{
heatMapGenome(
  obj,
  matrix = NULL,
  colors = NULL,
  legend = TRUE,
  removeNA = TRUE,
  colorMax = NULL,
  order = NULL
)
}
\arguments{
\item{obj}{Object containing the matrix to plot}

\item{matrix}{Name of matrix containing aggregated methylation values calculated over large genomic windows
See ?aggregateMatrix if needed.}

\item{colors}{Optional; specify color palette}

\item{legend}{Boolean indicating whether to plot the color legend}

\item{removeNA}{Boolean indicating whether to remove NA values as opposed to plotting them grey}

\item{colorMax}{Set upper bound on color scale. Everything exceeding this threshold will be plotted at the max value}

\item{order}{Arrange the order of groups}
}
\value{
A ggplot geom_tile object with colors indicating \% methylation over tiled windows and the gene of interest beneath
}
\description{
Plot genome-wide view of pseudobulked methylation levels aggregated across large genomic tiles
}
\examples{
\dontrun{
  heatMapGenome(obj = brain, matrix = "ch_100k_pct_type", colors = c("black", "red", "yellow"))
}
}
