% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{histograM}
\alias{histograM}
\title{histograM}
\usage{
histograM(
  obj,
  genes = NULL,
  matrix,
  colors = NULL,
  trackOverhang = 5000,
  arrowOverhang = 3000,
  ncol = length(genes),
  legend = TRUE,
  removeNA = TRUE,
  width = 500,
  trackScale = 1.5,
  colorMax = 100
)
}
\arguments{
\item{obj}{Object containing the matrix to plot}

\item{genes}{Gene list to plot}

\item{matrix}{Matrix contained in the genomeMatrices slot; pct matrix output of calcSmoothedWindows}

\item{colors}{Optional list of colors to include}

\item{trackOverhang}{Number of base pairs to extend beyond the gene}

\item{arrowOverhang}{Number of base pairs the track arrow should extend beyond the gene}

\item{ncol}{Number of columns to distribute results across. Defaults to number of input genes}

\item{legend}{Boolean indicating whether to plot the color legend}

\item{removeNA}{Boolean indicating whether to remove NA values as opposed to plotting them grey}

\item{width}{Width of the tiles, which should correspond to the step size calculated with calcSmoothedWindows}
}
\value{
A ggplot geom_tile object with colors indicating \% methylation over 500 bp windows and the gene of interest beneath
}
\description{
Plot methylation levels over a gene body in histogram format
}
\examples{
histograM(obj, genes = c("SATB2", "BCL11B"), matrix = "cluster_cg_500_slidingwindows", legend = F)
}
