% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{histograM}
\alias{histograM}
\title{histograM}
\usage{
histograM(
  obj,
  genes = NULL,
  matrix,
  colors = NULL,
  trackOverhang = 5000,
  arrowOverhang = 3000,
  ncol = length(genes),
  legend = TRUE,
  removeNA = TRUE,
  trackScale = 1.5,
  arrowScale = 0.025,
  colorMax = 100,
  order = NULL,
  baseline = 0
)
}
\arguments{
\item{obj}{Object containing the matrix to plot}

\item{genes}{Gene list to plot}

\item{matrix}{Matrix contained in the genomeMatrices slot; pct matrix output of calcSmoothedWindows}

\item{colors}{Optional list of colors to include}

\item{trackOverhang}{Number of base pairs to extend beyond the gene}

\item{arrowOverhang}{Number of base pairs the track arrow should extend beyond the gene}

\item{ncol}{Number of columns to distribute results across. Defaults to number of input genes}

\item{legend}{Boolean indicating whether to plot the color legend}

\item{removeNA}{Boolean indicating whether to remove NA values as opposed to plotting them grey}

\item{trackScale}{Enables adjustment of gene track height}

\item{arrowScale}{Enables adjustment of gene directional arrow size}

\item{colorMax}{Set upper bound on color scale. Everything exceeding this threshold will be plotted at the max value}

\item{order}{Arrange the order of groups}

\item{baseline}{0 or "mean". Define whether the histogram baseline starts at 0 or the mean global \% methylation of the group.}
}
\value{
A ggplot geom_tile object with colors indicating \% methylation over tiled windows and the gene of interest beneath
}
\description{
Plot methylation levels over a gene body in histogram format
}
\examples{
histograM(brain, genes = "SATB2", matrix = "cg_type_tracks")
histograM(brain, genes = "SATB2", matrix = "ch_type_tracks", baseline = "mean", arrowScale = .03, trackScale = .5, colors = c("#ff4a74", "grey20", "#71e300"), colorMax = 12, trackOverhang = 50000, order = rev(c("Astro", "Oligo", "OPC", "Micro", "Inh_CGE", "Inh_MGE", "Exc_L1-3_CUX2", "Exc_L2-4_RORB", "Exc_L4-5_FOXP2", "Exc_L4-6_LRRK1", "Exc_L5-6_PDZRN4", "Exc_L6_TLE4")))
}
