% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{impute}
\alias{impute}
\title{impute}
\usage{
impute(obj, matrix, npca = 10, replaceNA = 0)
}
\arguments{
\item{obj}{Amethyst object to perform imputation on}

\item{matrix}{Name of the matrix contained in the genomeMatrices slot to perform imputation on}

\item{npca}{Number of principle components to use}

\item{replaceNA}{Rmagic can't accept NA values. Replace NA values with 0, 1, "mch_pct", or "mcg_pct".}
}
\value{
Returns a data frame of imputed values.
}
\description{
Impute values with the Rmagic package https://github.com/KrishnaswamyLab/MAGIC
Warning: we recommend using with caution. Ideally, any results would also be visible without imputation, and this
would be used to help smooth values.
}
\examples{
obj <- impute(obj, matrix = "gene_ch")
}
