% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{indexChr}
\alias{indexChr}
\title{indexChr}
\usage{
indexChr(obj, type, threads = 1)
}
\arguments{
\item{type}{What type of methylation to retrieve; i.e. mCH or mCG.}

\item{threads}{Optional threads parameter to use if parallelizing.}

\item{hdf5}{Path to the hdf5 file containing base-level read information
organized by methylation type and barcode.}
}
\value{
A list of data frames named by cell barcode. The data frames contain the
coordinates in the hdf5 file corresponding to each chromosome.
}
\description{
If the whole hdf5 file had to be searched for relevant reads
every time gene-specific methylation information was needed, most functions
would take minutes to hours to run. The indexing process catalogs the
coordinates for each specified gene in every cell beforehand so the relevant
subset can quickly be retrieved for downstream processes. mCG, mCH, and the
mCG status of the promoter must be indexed separately. Providing a gene
subset is recommended due to memory and time constraints.
}
