% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{makePalette}
\alias{makePalette}
\title{makePalette}
\usage{
makePalette(option, n)
}
\arguments{
\item{option}{Which palette to choose out of options 1-30. Use testPalette to see the colors in each.}

\item{n}{How many colors are needed in the palette.}
}
\value{
Returns a character vector of n hex codes.
}
\description{
Generates a color palette of any length
}
\examples{
pal <- makePalette(option = 1, n = 20)
}
