% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{makeRef}
\alias{makeRef}
\title{makeRef}
\usage{
makeRef(ref = "hg38", gtf = NULL, attributes = c("gene_name", "exon_number"))
}
\arguments{
\item{attributes}{}
}
\description{
Generate an annotation file from a gtf. Paths to mm10 and hg38 are hard-coded. Please provide the gtf if analyzing other species.
}
