% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R, R/helper_dt.R
\name{makeWindows}
\alias{makeWindows}
\title{makeWindows}
\usage{
makeWindows(
  obj,
  type,
  groupBy = NULL,
  genes = NULL,
  stepsize = FALSE,
  nmin = NULL,
  metric = "percent",
  threads = 1,
  bed = NULL
)

makeWindows(
  obj,
  type,
  groupBy = NULL,
  genes = NULL,
  stepsize = FALSE,
  nmin = NULL,
  metric = "percent",
  threads = 1,
  bed = NULL
)
}
\arguments{
\item{obj}{Object for which to calculate percent methylation windows}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene
body mCG, or promoter mCG.}

\item{groupBy}{Optional metadata parameter to group cells by. If included,
the function will calculate mean percent methylation of a given window over
the group.}

\item{genes}{If specified, the \% methylation will be calculated over the
indicated gene bodies.}

\item{stepsize}{If specified, the \% methylation will be calculated over
genomic windows of the indicated size.}

\item{nmin}{Optional. Minimum number of cytosines recorded in the window/gene
to be included in the output.}

\item{metric}{Calculate either percent, score, or ratio}

\item{threads}{Enables multithreading}
}
\value{
Returns a data frame of the calculated methylation windows with the genomic region or gene name as rows and cell barcode or group as columns

Returns a data frame with the genomic region or gene name as column
names, cell barcode or group as row names, and values as percent methylated.
}
\description{
Calculate methylation levels across a fixed genomic window or gene body.

Calculate percent methylation across a genomic window of given
size or a gene body. If desired, aggregate average percent methylation
according to a categorical value in the metadata.
}
\examples{
makeWindows(obj = obj, groupBy = "cluster_id", genes = c("SATB2", "GAD1"), type = "CH", metric = "percent", threads = 10)
makeWindows(obj = obj, stepsize = 100000, type = "CG", metric = "score")
\dontrun{
makeWindows(obj, groupBy = cluster_id, genes = c("TBR2", "TBR1"), nmin = 10, type = "CH")
makeWindows(obj, stepsize = 100000, type = "CH")
}
}
