% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{makeWindows}
\alias{makeWindows}
\title{makeWindows
Calculate percent methylation across a genomic window of given size or a gene body.
If desired, aggregate average percent methylation according to a categorical value in the metadata.}
\usage{
makeWindows(
  obj,
  groupBy = NULL,
  genes = NULL,
  stepsize = FALSE,
  nmin = NULL,
  type,
  metric = "percent",
  threads = 1,
  bed = NULL
)
}
\arguments{
\item{obj}{Object for which to calculate percent methylation windows}

\item{groupBy}{Optional metadata parameter to group cells by. If included, the function will calculate mean percent methylation of a given window over the group.}

\item{genes}{If specified, the \% methylation will be calculated over the indicated gene bodies.}

\item{stepsize}{If specified, the \% methylation will be calculated over genomic windows of the indicated size.}

\item{nmin}{Optional. Minimum number of cytosines recorded in the window/gene to be included in the output.}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG.}
}
\value{
Returns a data frame with the genomic region or gene name as column names, cell barcode or group as row names, and values as percent methylated.
}
\description{
makeWindows
Calculate percent methylation across a genomic window of given size or a gene body.
If desired, aggregate average percent methylation according to a categorical value in the metadata.
}
\examples{
makeWindows(obj, groupBy = cluster_id, genes = c("TBR2", "TBR1"), nmin = 10, type = "CH")
makeWindows(obj, stepsize = 100000, type = "CH")
}
