% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{markerHeatmap}
\alias{markerHeatmap}
\title{markerHeatmap
NEEDS UPDATING}
\usage{
markerHeatmap(
  obj,
  level = "CellClass",
  impute = TRUE,
  rownames = FALSE,
  type = "CH"
)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{level}{Hierarchy of reference to plot. i.e., "cell_class" plots "Exc"  "Inh"  "NonN"}

\item{impute}{Logical indicating whether to impute values using Rmagic}

\item{rownames}{Logical indicating whether to show row names on the heatmap}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG.}
}
\value{
Returns a heatmap of average methylation levels of key marker genes
}
\description{
Plot the methylation levels of sample groups across many genes at once
}
\examples{
markerHeatmap(obj = obj, level = "CellClass", impute = FALSE, rownames = FALSE, type = "CH")
}
