% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{regressCovBias}
\alias{regressCovBias}
\title{regressCovBias}
\usage{
regressCovBias(obj, reduction = "irlba")
}
\arguments{
\item{obj}{Amethyst object containing irlba matrix to regress coverage bias from}
}
\value{
Returns the residuals
}
\description{
Calculate the residuals of a logistic regression of cell coverage vs. each irlba dimension
}
\examples{
obj@reductions[["irlba_regressed"]] <- regressCovBias(obj)
}
