% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{regressCovBias}
\alias{regressCovBias}
\title{regressCovBias}
\usage{
regressCovBias(obj, reduction = "irlba", method = "lm")
}
\arguments{
\item{obj}{Amethyst object containing irlba matrix to regress coverage bias from}

\item{reduction}{Name of dimensionality reduction to apply function over}

\item{method}{Calculate residuals using "lm" or "gam" model}
}
\value{
Returns the residuals
}
\description{
Calculate the residuals of a logistic regression of cell coverage vs. each irlba dimension
}
\examples{
obj@reductions[["irlba_regressed"]] <- regressCovBias(obj)
}
