% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{runCluster}
\alias{runCluster}
\title{runCluster}
\usage{
runCluster(
  obj,
  k = 50,
  reduction = "irlba",
  method = "louvain",
  resolution = 0.001
)
}
\arguments{
\item{obj}{Amethyst object to perform clustering on}

\item{k}{integer; number of nearest neighbors. Option applies to louvain method only.}

\item{reduction}{Name of dimensionality reduction to calculate over}

\item{method}{Options are "louvain", which utilizes the Rphenograph https://github.com/JinmiaoChenLab/Rphenograph package; or "leiden", which utilizes the Trapnell lab's implementation of the Leiden algorithm https://github.com/cole-trapnell-lab/leidenbase}
}
\value{
Returns cluster membership vector to be appended to the metadata of an amethyst object
}
\description{
Determine cluster membership with Rphenograph https://github.com/JinmiaoChenLab/Rphenograph
}
\examples{
\dontrun{
  obj <- runCluster(obj = obj, k = 30, reduction = "irlba", method = "louvain", colname = "clusters_k30_irlba_louvain")
}
}
