% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{runCluster}
\alias{runCluster}
\title{runCluster}
\usage{
runCluster(obj, k_phenograph = 50, reduction = "irlba")
}
\arguments{
\item{obj}{Amethyst object to perform clustering on}

\item{k_phenograph}{integer; number of nearest neighbors}
}
\value{
Adds cluster membership to the metadata file of the amethyst object
}
\description{
Determine cluster membership with Rphenograph https://github.com/JinmiaoChenLab/Rphenograph
}
\examples{
obj <- runCluster(obj = obj, k_phenograph = 30)
}
