% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{runCluster}
\alias{runCluster}
\title{runCluster
Cluster dimensionality reduction with Rphenograph}
\usage{
runCluster(obj, k_phenograph = 50)
}
\arguments{
\item{k_phenograph}{}
}
\description{
runCluster
Cluster dimensionality reduction with Rphenograph
}
