% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runIrlba}
\alias{runIrlba}
\title{runIrlba}
\usage{
runIrlba(
  obj,
  genomeMatrices,
  dims,
  name = "irlba",
  replaceNA = rep(0, length(dims))
)
}
\arguments{
\item{obj}{Object for which to run irlba}

\item{genomeMatrices}{list of matrices in the genomeMatrices slot to use for irlba}

\item{dims}{list of how many dimensions to output for each matrix}

\item{replaceNA}{IRLBA can't accept NA values. Replace NA values with 0, 1, "mch_pct", or "mcg_pct".}
}
\value{
Returns a matrix of appended irlba dimensions as columns and cells as rows
}
\description{
Perform dimensionality reduction based on methylation levels over a matrix stored in the @genomeMatrices slot
}
\examples{
obj@reductions[["irlba"]] <- runIrlba(obj, genomeMatrices = c("ch_2M_pct", "ch_2M_score", "cg_2M_score"), dims = c(10, 10, 10))
}
