% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runUmap}
\alias{runUmap}
\title{runUmap}
\usage{
runUmap(
  obj,
  neighbors = 30,
  dist = 0.1,
  method = "euclidean",
  reduction = "irlba"
)
}
\arguments{
\item{obj}{Amethyst object for which to determine umap coordinates}

\item{neighbors}{Number of closest points to factor into projection calculations. A higher number will capture more global structure}

\item{dist}{Distance between point pairs}

\item{method}{Distance metric to utilize. Default is euclidean}

\item{reduction}{Character; the name of the dimensionality reduction to use from the object. Default is "irlba"}
}
\value{
A data frame of UMAP x and y coordinates labeled "dim_x" and "dim_y"
}
\description{
Perform dimension reduction with Uniform Manifold APproximation and Projection for Dimension Reduction
}
\examples{
\dontrun{
  obj@reductions[["umap"]] <- runUmap(obj, neighbors = 30, dist = 0.1, method = "euclidean", reduction = "irlba")
}
}
