% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{runUmap}
\alias{runUmap}
\title{runUmap
Perform dimension reduction with Uniform Manifold APproximation and Projection for Dimension Reduction}
\usage{
runUmap(obj, neighbors = 30, dist = 0.1, method = "euclidean")
}
\arguments{
\item{method}{}
}
\description{
runUmap
Perform dimension reduction with Uniform Manifold APproximation and Projection for Dimension Reduction
}
