% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{sampleComp}
\alias{sampleComp}
\title{sampleComp
Visualize the distribution of a metadata variable across groups of cells}
\usage{
sampleComp(obj, groupBy, colorBy, colors = NULL)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{groupBy}{What column in the metadata to group cells by. This will be on the x axis.}

\item{colorBy}{What column in the metadata to color the bar chart by.}
}
\value{
A bar graph showing the distribution of categorical metadata in user-defined groups of cells
}
\description{
sampleComp
Visualize the distribution of a metadata variable across groups of cells
}
\examples{
sampleComp(obj, groupBy = sample, colorBy = cluster_id)
}
