% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{umapFeature}
\alias{umapFeature}
\title{UMAP feature
Plot any feature in the metadata in UMAP space}
\usage{
umapFeature(obj, colorBy, colors = NULL, pointSize = 0.5)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{colorBy}{The metadata feature to plot in UMAP space}

\item{colors}{Optional color scale}

\item{pointSize}{Optional adjustment of point size}
}
\value{
Returns a ggplot graph plotting xy coordinates of cells colored according to the specified feature
}
\description{
UMAP feature
Plot any feature in the metadata in UMAP space
}
\examples{
umapFeature(obj, colorBy = cluster_id)

}
