% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{umapGeneM}
\alias{umapGeneM}
\title{umapGeneM}
\usage{
umapGeneM(
  obj,
  genes,
  type = NULL,
  metric = "percent",
  matrix = NULL,
  blend = FALSE,
  colors = NULL,
  squish = NULL,
  nrow = round(sqrt(length(genes)))
)
}
\arguments{
\item{obj}{The amethyst object to plot}

\item{genes}{List of genes to plot \% methylation in UMAP space}

\item{type}{What type of methylation to retrieve; i.e. gene body mCH, gene body mCG, or promoter mCG.}

\item{metric}{If matrix is not specified, what methylation parameter to calculate, i.e. "percent", "score", or "ratio"}

\item{matrix}{Pre-computed matrix to use in the genomeMatrices slot, if available. The type parameter is not needed if matrix is definied.}

\item{blend}{Logical indicating whether to blend methylation levels of two or three genes}

\item{colors}{Optional color gradient to include}

\item{nrow}{Number of rows to distribute graphs over}
}
\value{
Returns a plot of methylation over given genomic region in relation to UMAP
}
\description{
Plot cumulative \% methylation over a gene body on umap
}
\examples{
umapGeneM(obj, genes = c("GAD1", "SATB2"), type = "CH", metric = "percent", colors = c("black", "turquoise", "gold", "red"), blend = F)
umapGeneM(both, genes = c("GAD1", "SATB2"), matrix = "gene_ch", blend = T, nrow = 1)
}
